+++ CHANGELOG.md
@@ -11,7 +11,14 @@ This project does *not* adhere to [Semantic Versioning](https://semver.org).
 ## [Unreleased](https://github.com/dsa-ou/algoesup/compare/v0.4...HEAD)
 These changes are in the GitHub repository but not on [PyPI](https://pypi.org/project/algoesup).
 
-Nothing yet.
+<!-- Nothing yet. -->
+
+### Fixed
+- don't show Ruff's `invalid-syntax` messages
+
+### Changed
+- show `allowed` warnings for the current code cell
+- show warnings for `-f` and for no `-m` when `allowed` is activated
 
 ## [0.4](https://github.com/dsa-ou/algoesup/compare/v0.3.1...v0.4) - 2025-08-11
 ### Added